feat: add route /c-by-id/<projectId> and redirect /projects-by-id/<projectId>
